+++ NEWS.md
@@ -1,5 +1,8 @@
 # cellKey 0.18.1
 - bugfix: fallback to use a single core on windows-machines, fixing [`issue #131`](https://github.com/sdcTools/UserSupport/issues/131)
+- updating dependencies and required versions
+- fix vignette due to updates in `ptable`-pkg
+- simplify examples by using examplary ptable from `ptable`-pkg using `ptable::pt_ex_cnts()` and `ptable::pt_ex_nums()`
 
 # cellKey 0.18.0
 - allow to save perturbation-schemes for different variables in `params_cnts_set()` and `params_nums_set()`